Update directory and graphql schema structure
