+++ backend/microservices/registry/schemas/schema-builder.js
@@ -3,19 +3,19 @@
 // Require modules.
 const { buildSchema } = require('graphql');
 const {concatenateSchemas} = require('./utils');
-const { userSchemaDefinition } = require('./user-schema');
-const { rootSchemaDefinition } = require('./root-schema');
 const { queries } = require('./queries');
 const { mutations } = require('./mutations');
-const { UserRegistrationInput } = require('./inputs');
+const { UserRegistrationInput, UserLoginInput } = require('./inputs');
+const { userSchemaDefinition, rootSchemaDefinition, authenticationDataDefinition } = require('./typedefs');
 
 // Concatenate schemas, queries and mutations.
-const concatenatedSchemas = concatenateSchemas(
-    userSchemaDefinition, rootSchemaDefinition, queries, mutations, UserRegistrationInput
+const concatenatedSchema = concatenateSchemas(
+    userSchemaDefinition, rootSchemaDefinition, queries, mutations, UserRegistrationInput,
+    authenticationDataDefinition, UserLoginInput
 );
 
 // Build graphql schema.
-const graphqlSchema = buildSchema(concatenatedSchemas);
+const graphqlSchema = buildSchema(concatenatedSchema);
 
 // Export schema.
 exports.graphqlSchema = graphqlSchema;
